requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
openai>=1.54.0
azure-identity>=1.15.0
streamlit>=1.39.0
neo4j>=5.26.0
py2neo>=2021.2.4
networkx>=3.4.0
sentence-transformers>=3.2.0

# Graph Visualization
streamlit-agraph>=0.0.45
plotly>=5.17.0
pyvis>=0.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0
jsonschema>=4.19.0
